Set up userspace for common keymap elements
