+++ keyboards/hhkb/keymaps/tominabox1/config.h
@@ -36,20 +36,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #endif
 #define MATRIX_COLS 8
 
-/* Fast typing */
-#define IGNORE_MOD_TAP_INTERRUPT
-//#define PERMISSIVE_HOLD
-#define TAPPING_TERM_PER_KEY
-//#define TAPPING_FORCE_HOLD
-#define TAPPING_TERM 300
-//#define RETRO_TAPPING
-
-/* number of backlight levels */
-#define BACKLIGHT_LEVELS 3
-
-/* Set 0 if debouncing isn't needed */
-#define DEBOUNCE    5
-
 /* Mechanical locking support. Use KC_LCAP, KC_LNUM or KC_LSCR instead in keymap */
 //#define LOCKING_SUPPORT_ENABLE
 /* Locking resynchronize hack */
@@ -96,24 +82,4 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #endif
 
-/*
- * Feature disable options
- *  These options are also useful to firmware size reduction.
-*/
-
-/* disable debug print */
-//#define NO_DEBUG
-
-/* disable print */
-//#define NO_PRINT
-
-/* disable action features */
-//#define NO_ACTION_LAYER
-//#define NO_ACTION_TAPPING
-//#define NO_ACTION_ONESHOT
-//#define NO_ACTION_MACRO
-//#define NO_ACTION_FUNCTION
-
 #endif
-
-
