--- keyboards/lazydesigners/dimple/keymaps/tominabox1/config.h
@@ -1,79 +1,0 @@
-/*
-Copyright 2019 Erovia
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 2 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#pragma once
-
-#include "config_common.h"
-
-/* USB Device descriptor parameter */
-#define VENDOR_ID       0xFEED
-#define PRODUCT_ID      0x0040
-#define DEVICE_VER      0x0001
-#define MANUFACTURER    LazyDesigners
-#define PRODUCT         Dimple
-#define DESCRIPTION     A 40% keyboard
-
-
-/* key matrix size */
-#define MATRIX_ROWS 4
-#define MATRIX_COLS 12
-
-/* key matrix pins */
-#define MATRIX_ROW_PINS { D0, D1, D2, D3 }
-#define MATRIX_COL_PINS { B0, B1, B2, B3, D4, D6, D7, B4, B5, B6, C6, C7 }
-#define UNUSED_PINS
-
-/* COL2ROW or ROW2COL */
-#define DIODE_DIRECTION COL2ROW
-
-/* Set 0 if debouncing isn't needed */
-#define DEBOUNCING_DELAY 5
-
-/* Mechanical locking support. Use KC_LCAP, KC_LNUM or KC_LSCR instead in keymap */
-#define LOCKING_SUPPORT_ENABLE
-
-/* Locking resynchronize hack */
-#define LOCKING_RESYNC_ENABLE
-
-/* Fast typing permissive hold */
-#define IGNORE_MOD_TAP_INTERRUPT
-#define PERMISSIVE_HOLD
-#define TAPPING_TERM_PER_KEY
-//#define TAPPING_FORCE_HOLD
-#define TAPPING_TERM 200
-//#define RETRO_TAPPING
-
-/* RBG underglow */
-#define RGB_DI_PIN B7
-#ifdef RGB_DI_PIN
-	#define RGBLIGHT_ANIMATIONS
-	#define RGBLIGHT_SLEEP
-	#define RGBLED_NUM 50
-	/* #define RGBLIGHT_HUE_STEP 8 */
-	/* #define RGBLIGHT_SAT_STEP 8 */
-	/* #define RGBLIGHT_VAL_STEP 8 */
-#   define RGBLIGHT_HUE_STEP 8
-#   define RGBLIGHT_SAT_STEP 8
-#   define RGBLIGHT_VAL_STEP 8
-#   define RGBLIGHT_LIMIT_VAL 100
-#endif
-
-/* CapsLock LED */
-#define BACKLIGHT_PIN E6
-#ifdef BACKLIGHT_PIN
-	#define BACKLIGHT_LEVELS 6
-#endif
