+++ keyboards/lazydesigners/dimple/keymaps/tominabox1/keymap.c
@@ -14,62 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include QMK_KEYBOARD_H
-#define USB_MAX_POWER_CONSUMPTION 900
-
-#define KC_SFT_CPS MT(MOD_LSFT, KC_CAPS) // Act as Shift on hold and as CapsLock on tap
-#define SPC_LOW LT(_LOWER, KC_ENT) // Left space on tap, LOWER on hold
-#define SPC_UPR LT(_RAISE, KC_SPC) // Left space on tap, UPPER on hold
-
-enum custom_layers {
-	_QWERTY,
-	_LOWER,
-	_RAISE,
-	_ADJUST,
-  _FKEY
-};
-
-enum {
-    KC_EML = 0,
-    TD_SFT_CPS,
-};
-
-void dance_cln_finished (qk_tap_dance_state_t *state, void *user_data) {
-    if (state->count == 1) {
-        register_code (KC_RSFT);
-        register_code (KC_2);
-    } else {
-        SEND_STRING("tom.campie@gmail.com");
-    }
-}
-
-void dance_cln_reset (qk_tap_dance_state_t *state, void *user_data) {
-    if (state->count == 1) {
-        unregister_code (KC_RSFT);
-        unregister_code (KC_2);
-    } else {
-    }
-}
-
-//All tap dance functions would go here. Only showing this one.
-qk_tap_dance_action_t tap_dance_actions[] = {
-    [KC_EML] = ACTION_TAP_DANCE_FN_ADVANCED (NULL, dance_cln_finished, dance_cln_reset),
-    //[TD_SFT_CPS] = ACTION_TAP_DANCE_DOUBLE(KC_CAPS, KC_LSFT)
-};
-
-uint16_t get_tapping_term(uint16_t keycode) {
-    switch (keycode) {
-       case KC_SFT_CPS:
-           return 150;
-//        case SPC_UPR:
-//            return 50;
-        case SPC_LOW:
-            return 150;
-////        case TD(KC_EML):
-////            return 50;
-        default:
-            return TAPPING_TERM;
-    }
-};
+#include "tominabox1.h"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
 
@@ -85,10 +30,10 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
  *      `-----------------------------------------------------'
  */
 
-	[_QWERTY] = LAYOUT(
+	[_BASE] = LAYOUT(
     KC_GESC, KC_Q,    KC_W,    KC_F,    KC_P,    KC_G,    KC_J,    KC_L,    KC_U,    KC_Y,    KC_SCLN,    KC_BSPC,
     LCTL_T(KC_TAB),  KC_A,    KC_R,    KC_S,    KC_T,    KC_D,    KC_H,    KC_N,    KC_E,    KC_I,    KC_O, KC_QUOT,
-		KC_SFT_CPS, KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,    KC_K,    KC_M,    KC_COMM, KC_UP,   KC_DOT,
+		TD(TD_SFT_CPS), KC_Z,    KC_X,    KC_C,    KC_V,    KC_B,    KC_K,    KC_M,    KC_COMM, KC_UP,   KC_DOT,
 		KC_LCTL, KC_LGUI, KC_LALT, SPC_LOW,  SPC_UPR,  LT(_FKEY, KC_LEFT), KC_DOWN, KC_RGHT
 	),
 
@@ -105,9 +50,9 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
  */
 
 	[_LOWER] = LAYOUT(
-		KC_TILD, KC_EXLM, TD(KC_EML),   KC_HASH, KC_DLR,  KC_PERC, KC_CIRC, KC_AMPR, KC_ASTR, KC_LPRN, KC_RPRN, KC_DEL,
+		KC_GESC, KC_EXLM, TD(KC_EMAIL),   KC_HASH, KC_DLR,  KC_PERC, KC_CIRC, KC_AMPR, KC_ASTR, KC_LPRN, KC_RPRN, KC_DEL,
 		KC_DEL,  KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_UNDS, KC_PLUS, KC_LCBR, KC_RCBR, KC_PIPE,
-		KC_TRNS,   KC_NO,   KC_NO,   KC_NO,   KC_NO,  KC_NO,  KC_NO,  KC_NO,   KC_NO, KC_PGUP, KC_QUES,
+		TD(TD_SFT_CPS),   KC_NO,   KC_NO,   KC_NO,   KC_NO,  KC_NO,  KC_NO,  KC_NO,   KC_NO, KC_PGUP, KC_QUES,
 		LCA(KC_DEL),   KC_NO,   KC_NO,   KC_TRNS, KC_TRNS, KC_HOME, KC_PGDN, KC_END
 	),
 
@@ -124,7 +69,7 @@ const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
  */
 
 	[_RAISE] = LAYOUT(
-    KC_GRV,  KC_1,    KC_2,    KC_3,    KC_4,    KC_5,    KC_6,    KC_7,    KC_8,    KC_9,    KC_0,    KC_BSPC,
+    KC_GESC,  KC_1,    KC_2,    KC_3,    KC_4,    KC_5,    KC_6,    KC_7,    KC_8,    KC_9,    KC_0,    KC_BSPC,
 		KC_NO,  KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_MINS, KC_EQL,  KC_LBRC, KC_RBRC, KC_BSLS,
 		KC_NO,   KC_NO,   KC_NO,   KC_NO,   KC_NO,  KC_NO,  KC_NO,  KC_NO,   KC_NO,   KC_DOT,   KC_SLSH,
 		KC_VOLD, KC_MUTE, KC_VOLU, KC_TRNS, KC_TRNS, KC_TRNS,   KC_NO,   KC_NO
@@ -174,24 +119,4 @@ if (IS_LED_ON(usb_led, USB_LED_CAPS_LOCK)) {
 	} else {
 	dimple_led_on();
 	}
-}
-
-
-uint32_t layer_state_set_user(uint32_t state) {
-    state = update_tri_layer_state(state, _LOWER, _RAISE, _ADJUST);
-    switch (biton32(state)) {
-        case _LOWER:
-            rgblight_sethsv_noeeprom(HSV_BLUE);
-            break;
-        case _RAISE:
-            rgblight_sethsv_noeeprom(HSV_GREEN);
-            break;
-        case _ADJUST:
-            rgblight_sethsv_noeeprom(HSV_CYAN);
-            break;
-        default:
-            rgblight_sethsv_noeeprom(50,140,201);
-            break;
-    }
-    return state;
 }
