--- keyboards/lazydesigners/dimple/keymaps/tominabox1/readme.md
@@ -1,60 +1,0 @@
-# The default keymap for Dimple
-
-**Features**
-
-* Shift acts as CapsLock when tapped  
-* Left space: space on tap, Lower layer on hold  
-* Right space: space on tap, Raise layer on hold  
-* Both space: Adjust layer on hold  
-
-## QWERTY (Normal) Layer
-```
- ,---------------------------------------------------------------.
- |Esc | Q  | W  | E  | R  | T  | Y  | U  | I  | O  | P  |  Bspc  |
- |---------------------------------------------------------------|
- | Tab  | A  | S  | D  | F  | G  | H  | J  | K  | L  | '  |Enter |
- |---------------------------------------------------------------|
- | Shift  | Z  | X  | C  | V  | B  | N  | M  | ,  | Up |    .    |
- |---------------------------------------------------------------|
-      |Ctrl|Gui |Alt | Spc/Lwr |   Spc/Rse   |Left|Down|Rght|
-      `-----------------------------------------------------'
-```
-
-## Lower Layer
-```
- ,---------------------------------------------------------------.
- | ~  | !  | @  | #  | $  | %  | ^  | &  | *  | (  | )  |        |
- |---------------------------------------------------------------|
- | Del  | F1 | F2 | F3 | F4 | F5 | F6 | _  | +  | {  | }  |  |   |
- |---------------------------------------------------------------|
- |        | F7 | F8 | F9 |F10 |F11 |F12 |    | ;  |PgUp|   /     |
- |---------------------------------------------------------------|
-      |    |    |    |         |             |Home|PgDn|End |
-      `-----------------------------------------------------'
-```
-
-## Raise Layer
-```
- ,---------------------------------------------------------------.
- | `  | 1  | 2  | 3  | 4  | 5  | 6  | 7  | 8  | 9  | 0  |        |
- |---------------------------------------------------------------|
- | Ins  | F1 | F2 | F3 | F4 | F5 | F6 | -  | =  | [  | ]  |  \   |
- |---------------------------------------------------------------|
- |        | F7 | F8 | F9 |F10 |F11 |F12  |    |    |   |         |
- |---------------------------------------------------------------|
-      |VolD|Mute|VolU|         |             |    |    |    |
-      `-----------------------------------------------------'
-```
-
-## Adjust Layer
-```
-,---------------------------------------------------------------.
- |EEPR|RST |    |    |    |    |    |    |    |    |    |        |
- |---------------------------------------------------------------|
- |      |    |    |    |    |    |    |    |    |    |    |      |
- |---------------------------------------------------------------|
- |        |    |    |    |    |    |     |    |    |   |         |
- |---------------------------------------------------------------|
-      |RGB-|RGB |RGB+|         |             |    |    |    |
-      `-----------------------------------------------------'
-```
