+++ users/tominabox1/config.h
@@ -1,14 +1,11 @@
 #pragma once
 
-// place overrides here
-/* Fast typing permissive hold */
+#define TAPPING_TERM 200
 #define IGNORE_MOD_TAP_INTERRUPT
 #define PERMISSIVE_HOLD
 #define TAPPING_TERM_PER_KEY
-//#define TAPPING_FORCE_HOLD
-#define TAPPING_TERM 200
-//#define RETRO_TAPPING
 
 /* Set 0 if debouncing isn't needed */
 #define DEBOUNCING_DELAY 5
+
