+++ users/tominabox1/rules.mk
@@ -1,0 +1,8 @@
+# Build Options
+#   change to "no" to disable the options, or define them in the Makefile in
+#   the appropriate keymap folder that will get included automatically
+#
+
+TAP_DANCE_ENABLE = yes
+
+SRC += tominabox1.c
